Updating vim plugins and some custom stuff
* Plugins added:
  - vim-airline
  - vim-airline-themes
  - vim-fugitive
  - ctrlpvim
  - tagbar
  - rst-robotframework-syntax-vim
  - vim-autopep8
* Fixing Doxygen configuration
* Customizing airline
* Configuring autopep8

+++ install_config.sh
@@ -3,18 +3,22 @@ CURDIR=`pwd`
 
 # create backups, for blind sighted users that doesnt even look what
 # scripts does
-if [ -f ~/.vimrc ]; then
+if [ -f ~/.vimrc ] && [ ! -L ~/.vimrc ]; then
 	mv -f ~/.vimrc ~/.vimrc.bk
 fi
-if [ -f ~/.gvimrc ]; then
+if [ -f ~/.gvimrc ] && [ ! -L ~/.gvimrc ]; then
 	mv -f ~/.gvimrc ~/.gvimrc.bk
 fi
-if [ -d ~/.vim ]; then
+if [ -d ~/.vim ] && [ ! -L ~/.vim ]; then
 	mv -f ~/.vim ~/.vim.bk
+fi
+if [ -d ~/.config/nvim ] && [ ! -L ~/.config/nvim ]; then
+	mv -f ~/.config/nvim ~/.config/nvim.bk
 fi
 
 # create simlinks for the current dir, so our configuration is updated
 # if its modified in the git repository
-ln -sf ${CURDIR}/vimrc ~/.vimrc
+ln -sf ${CURDIR}/vim/init.rc ~/.vimrc
 ln -sf ${CURDIR}/gvimrc ~/.gvimrc
 ln -sf ${CURDIR}/vim ~/.vim
+ln -sf ${CURDIR}/vim ~/.config/nvim
